blog pages are now built using ISR
