+++ pages/[page].tsx
@@ -1,10 +1,13 @@
 import { GetStaticPaths } from 'next';
-import BlogPost from '../components/BlogPost';
-import RichPage from '../components/RichPage';
+
 import { CMS, NAVIGATION } from '../constants';
 import { CmsApi, generateLinkMeta, unslugify } from '../services/cms';
 import { SideMenuItem } from '../state/navigation';
 import { IPost, ISplitPage, isPost } from '../types/cms';
+import { isLocal } from '../utils/links';
+
+import BlogPost from '../components/BlogPost';
+import RichPage from '../components/RichPage';
 
 interface IPath {
   params: { page: string };
@@ -15,11 +18,13 @@ export const getStaticPaths: GetStaticPaths = async () => {
   // Hardcoded in navigation constants.
   // Contentful can edit entries but cannot add/remove
   // without touching code.
-  const navigationPaths: IPath[] = Object.values(
-    NAVIGATION.SIDE_MENU_ITEMS,
-  ).map(item => ({
-    params: { page: item.href },
-  }));
+  const navigationPaths: IPath[] = Object.values(NAVIGATION.SIDE_MENU_ITEMS)
+    .filter(item => {
+      return item.hasOwnRoute === undefined && isLocal(item.href);
+    })
+    .map(item => ({
+      params: { page: item.href.slice(1) },
+    }));
 
   const cms = new CmsApi();
   let posts: IPost[] = [];
@@ -47,6 +52,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
 };
 
 export async function getStaticProps({ params }) {
+  console.log(`Building %c${params.page} page`, 'color: purple;');
   const href = params?.page ?? '';
   const id = unslugify(String(href));
 
@@ -54,7 +60,7 @@ export async function getStaticProps({ params }) {
     const cms = new CmsApi();
     let page: ISplitPage | IPost;
     if (SideMenuItem[id]) {
-      page = await cms.fetchPageById(SideMenuItem[id] ?? '');
+      page = await cms.fetchPageById(SideMenuItem[id]);
     } else {
       page = await cms.fetchEntryBySlug(href, 'post');
       // embedded links in post body need metadata for preview
